Douglaspeucker (#7)
* Basic Planar fns, Simplifier interface

Added some handy planar functions, and defined an Interface for
Simplifers. This way we can add and change out simplifiers easily.

* Added doublaspeucker simplification algo.

* Added reference to the formula for PerpendicularDisance.

* Updated according to reviews.

Moved spherical functions to the spherical namespace.
Added references to docs used for algos.

* Removed haversine test for simplify. Because not sure if it is working.

* [issue #7] Pull out Douglaspeucker Algo to the geom package.

Removed the spherical version as we don't need it and there may be an
issue with the distance formula that was being used, and got pulled.

Improved test scaffolding, with the aim to add more tests.