Reimplemented Constrained Delaunay on top of Delaunay
The constraineddelaunay package had some bugs that were diffcult
to track down as it is trying to handle the general case. For the
Hitmap based makevalid we don't have to handle the general case.
We can assume that all segments will not intersect except at their
end points.

The original routines returned basic geom types, instead of Line
and Triangles; which then had to be converted back and forth. This
one now takes a set of points and segments (instead of a geometry) and returns
a set of Triangles.

Added test cases and helper methods.