Removed unneeded packages from triangulation code.
When we are ready to work on the setdiff and a generic
constraineddelaunay triangulation we can revert this commit.

This is more to make sure that we have good test that are in a
passing state. Currently several of tests for triangulate and
makevalid/setdiff fail, as that code is still being developed.

From this package it turned out that only package that was needed
was the the quadedge package. So, this commit removed the other
packages that were brought in from the merge.